requirements.txt
# onto_rag/requirements.txt

# Core Ontology Processing
rdflib>=6.0.0

# Lexical Indexing
whoosh>=2.7.0

# Embeddings & Semantic Search (optional for compare_app)
sentence-transformers>=2.2.0; python_version < "3.13"
faiss-cpu>=1.7.0

# Prompt Templating
Jinja2>=3.0.0

# LLM SDKs (optional; only some apps use these)
google-genai>=1.0.0  # Provides 'from google import genai'

# Streamlit Cloud currently offers ollama<=0.6.0 on PyPI. Newer versions are
# unavailable there, which caused deploy failures. Use a compatible range.
ollama>=0.5.0,<0.7

# Utilities
numpy>=1.21.0
pandas>=1.3.0
python-dotenv>=0.19.0
tqdm>=4.60.0
streamlit-aggrid

# Testing (optional, but good for development)
pytest>=6.2.0

# Local LLM / Hugging Face (optional; excluded on Python 3.13+ for compatibility)
transformers>=4.39.0; python_version < "3.13"
torch>=2.1.0; python_version < "3.13"
accelerate>=0.28.0; python_version < "3.13"
bitsandbytes>=0.42.0; python_version < "3.13"  # For 8-bit/4-bit quantization

# Analysis / visualization extras
scikit-learn>=1.4.0
pyarrow>=14.0.0
plotly>=5.18.0
